Change the way how unexpected content is handled inside an expression
Before this commit it was allowed to use one escaped XML tag inside an expression. But this concept doesn't work for multiple XML tags, comments, CDATA sections or processing instructions. With this comment it isn't allowed to use any of the previous mentioned elements inside an expression. The parser throws now an SyntaxErrorException if this case occurs.